Revert "Make java_sdk_library dependencies explicit"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88e3025d-9532-4235-8f20-0c41fadd3c65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7eecde12ad4f8c1e0e2f80eb027032c53326ce4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com