Differentiate AdapterHolder types, pager indexes
The PRIMARY, WORK, and SEARCH values defined in AdapterHolder
classes now solely act as AdapterHolder types, rather than doubling
as indexes of tabs in the pager view. All cases in which these were
previously used as page indexes, or in which indexes were assumed
to correspond with types, have been replaced with the new PRIMARY_PAGE
or WORK_PAGE constqnts, or with calls to the new getPageForType and
getTypeForPage methods, as an initial step towards allowing multiple
pages of a given type (i.e. multiple WORK pages).

Part of this change means no longer pretending that SEARCH is an
actual page with an index, instead preferring to use methods like
getCurrentAdapterHolderType(). Where this was not possible because
pager indexes were directly involved, call sites were modified to
check against isSearching() directly.

Additionally, because adapter holders include a SEARCH entry that
pager views do not, we add and use some simple functions for
converting to and from AdapterHolder and pager indexes.

Bug: 379422740
Change-Id: If51bc1085e24f1c318d2b4b278c0973137a6873f