Fix a crash when searching widgets
There was an IllegalArgumentException when sorting widget search
results: "Comparison method violates its general contract!"
This was apparently due to inconsistent comparisons caused by the way
the comparator was attempting to prioritize the current user.
This change addresses that problem and adds a final lowest-priority
sort based on the user ID, which will typically show first-added
work profiles first.

Bug: 379422740
Change-Id: I6fc71dbda2d3f625a23e59aea8dd94125cfa6166