Fix click to launch the APP does not respond
According to the View of the call stack, quick click, the Launcher calls the cancelCurrentPageLongPress API, lead to View and remove the Tap, to the normal click event can't response.
The Launcher of cancelCurrentPageLongPress is called, because "if (absDeltaX & gt; mTouchSlop || absDeltaY &gt;  mTouchSlop) {" has been judged as true.
Tried to increase the value of mTouchSlop, but this problem basically didn't occur.

Bug: 430711008
Test: m
    1. Start the APP
    2. Press the home key and then immediately launch another app
    3. Repeat steps 1 and 2

Change-Id: I1019e002bd48b24ac0f185890dc6aaec2c43345b