Optimize desktop left and right sliding
Before starting, determine whether the distance between the touch starting point and the current touch point meets the scrolling conditions. The first frame that meets the scrolling conditions does not perform a scrolling action (only the state is updated, and scrolling starts from the next frame). The touch point and the finger starting point have already generated a sliding distance.

Bug: 422656135
Test: View trace
Change-Id: I5bc70256658f7203490c0e94dd915ecf392ee838