Proto test: Route item clicks through open folders in BaseDraggingActivity
This change modifies `BaseDraggingActivity` to allow an open `Folder` to intercept and handle click events on items within it.

`BaseDraggingActivity` now includes an abstract `getOpenFolder()` method. `getItemOnClickListener()` in `BaseDraggingActivity` will return the open folder's listener if a folder is open. `Launcher` extends `BaseDraggingActivity` and implements `getOpenFolder()` to return the currently open `mFolder`. `Folder` now implements `View.OnClickListener` and delegates the click handling to `ItemClickHandler.INSTANCE`. This ensures that clicks on items inside an open folder are first routed through the `Folder` instance.

Change-Id: I559bb30871b1b4bbe5d2fb7473f353e69cf63635